Extract Puzzletron wizard common helpers
Keep wizard.py as a compatibility facade so existing section imports and call sites retain exact behavior while later splits can share the foundation.
